updated layout
